style: fix C4244/C4267 in NoteDataWithScoring/NoteDisplay/ScreenGameplay/OptionsBinding (item 2)
Continuing the item-2 sweep (187 -> 158 remaining unique sites). All
casts document existing intentional behavior, no logic change --
characterization-verified (sm_tests unchanged at 5966/226).

- NoteDataWithScoring.cpp (11 sites, SS5-adjacent NoteData* per
  AGENTS.md 5 -- characterization only): RadarValues is float-backed
  but conceptually holds int counts in some slots. Reading counts back
  (out[RadarCategory_Notes] etc. -> int) and writing scoring counters
  into it (state.taps_hit etc. -> out[rc]) both cast explicitly now.
- NoteDisplay.cpp (9 sites): degrees-as-double constants (PI_180 /
  PI_180R) multiplied into a float rotation; the same IsOnScreen()
  float-members-into-int-parameters pattern NoteField.cpp had; a
  pointer-difference (v - buf) assigned/returned as int x2.
- ScreenGameplay.cpp (9 sites) + OptionsBinding.h's shared
  FLOAT_TABLE_INTERFACE macro (2 of the 9): SafeFArg()'s int return
  written into float margins[][2]; an iterator-diff into an int index;
  an enum subtraction into an int x2; lua_pushnumber(size_t) (should be
  lua_Number) at its direct call site and, fixed once at the macro
  definition so every other FLOAT_TABLE_INTERFACE user benefits too,
  size()/n+1 into lua_createtable/lua_rawseti's int parameters.

Verified: sm_tests 5966/226, ctest 100%, Release build + --SelfTest
green, with /wd4244/4267 removed AND restored (158 sites remain across
72 files, so the suppression stays for now).

+++ src/NoteDataWithScoring.cpp
@@ -352,9 +352,9 @@ void NoteDataWithScoring::GetActualRadarValues(const NoteData &in,
 
 	// ScreenGameplay passes in the RadarValues that were calculated by
 	// NoteDataUtil::CalculateRadarValues, so those are reused here. -Kyz
-	int note_count= out[RadarCategory_Notes];
-	int jump_count= out[RadarCategory_Jumps];
-	int hold_count= out[RadarCategory_Holds];
+	int note_count= static_cast<int>( out[RadarCategory_Notes] );
+	int jump_count= static_cast<int>( out[RadarCategory_Jumps] );
+	int hold_count= static_cast<int>( out[RadarCategory_Holds] );
 	float hittable_steps_length= std::max(0.0f,
 		timing->GetElapsedTimeFromBeat(NoteRowToBeat(last_hittable_row)) -
 		timing->GetElapsedTimeFromBeat(NoteRowToBeat(first_hittable_row)));
@@ -380,31 +380,31 @@ void NoteDataWithScoring::GetActualRadarValues(const NoteData &in,
 				out[rc]= GetActualChaosRadarValue(in, song_seconds, pss);
 				break;
 			case RadarCategory_TapsAndHolds:
-				out[rc]= state.taps_hit;
+				out[rc]= static_cast<float>( state.taps_hit );
 				break;
 			case RadarCategory_Jumps:
-				out[rc]= state.jumps_hit;
+				out[rc]= static_cast<float>( state.jumps_hit );
 				break;
 			case RadarCategory_Holds:
-				out[rc]= state.holds_held;
+				out[rc]= static_cast<float>( state.holds_held );
 				break;
 			case RadarCategory_Mines:
-				out[rc]= state.mines_avoided;
+				out[rc]= static_cast<float>( state.mines_avoided );
 				break;
 			case RadarCategory_Hands:
-				out[rc]= state.hands_hit;
+				out[rc]= static_cast<float>( state.hands_hit );
 				break;
 			case RadarCategory_Rolls:
-				out[rc]= state.rolls_held;
+				out[rc]= static_cast<float>( state.rolls_held );
 				break;
 			case RadarCategory_Lifts:
-				out[rc]= state.lifts_hit;
+				out[rc]= static_cast<float>( state.lifts_hit );
 				break;
 			case RadarCategory_Fakes:
 				out[rc]= out[rc];
 				break;
 			case RadarCategory_Notes:
-				out[rc]= state.notes_hit;
+				out[rc]= static_cast<float>( state.notes_hit );
 				break;
 			DEFAULT_FAIL(rc);
 		}
